Small tweaks with icons
