+++ App/Screens/Search/GpsItem/GpsItem.tsx
@@ -1,0 +1,39 @@
+// Sh**t! I Smoke
+// Copyright (C) 2018-2019  Marcelo S. Coelho, Amaury Martiny
+
+// Sh**t! I Smoke is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Sh**t! I Smoke is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Sh**t! I Smoke.  If not, see <http://www.gnu.org/licenses/>.
+
+import React, { useContext } from 'react';
+
+import { Item } from '../Item';
+import { i18n } from '../../../localization';
+import { CurrentLocationContext, GpsLocationContext } from '../../../stores';
+
+export function GpsItem () {
+  const { setCurrentLocation } = useContext(CurrentLocationContext);
+  const gps = useContext(GpsLocationContext);
+
+  const handleClick = () => {
+    setCurrentLocation(gps);
+  };
+
+  return (
+    <Item
+      description={i18n.t('search_current_location')}
+      icon="gps"
+      onPress={handleClick}
+      title={'HELLO'}
+    />
+  );
+}
