+++ App/localization/languages/en.json
@@ -21,7 +21,7 @@
   "home_common_oh": "Oh",
   "home_common_cigarette": "Cigarette",
   "home_common_cigarettes": "Cigarettes",
-  "home_smoked_cigarette_title": "{{swearWord}}! {{presentPast}} <{{cigarettes}} {{singularPlural}}>",
+  "home_smoked_cigarette_title": "{{swearWord}}! {{presentPast}}\n<{{cigarettes}} {{singularPlural}}>",
   "home_smoked_there": "there today",
   "home_swear_word_dang": "Dang",
   "home_swear_word_shoot": "Shoot",
@@ -33,6 +33,7 @@
   "loading_title_cough": "Cough",
   "loading_title_loading": "Loading",
   "search_header_input_placeholder": "Search for a city or address",
+  "search_current_location": "Your current location",
   "nav_btn_back": "Back",
   "details_air_quality_station_marker": "Air Quality Station",
   "details_your_position_marker": "Your position",
